requirements.txt
mysql-connector-python==8.1.0
pika==1.3.2
python-dotenv==1.0.0
pymongo==4.5.0